[fix] 적금이자수정, detail조회 userid로 조회하도록 수정

+++ backend/src/main/java/com/example/donzoom/dto/savings/response/SavingsResponseDto.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @Builder
 public class SavingsResponseDto {
-  private final int principal;   // 원금
-  private final int interest;    // 예상 이자
-  private final int totalAmount; // 총 예상 금액
+  private final double principal;   // 원금
+  private final double interest;    // 예상 이자
+  private final double totalAmount; // 총 예상 금액
 }
